Configurable event mail footer
Allow a configurable event mail footer setting (suite or global) to be
added to all event mails. The new setting, although specified at the
suite level, will apply to both suite events and task events.

Move task event handler logic from `cylc.task_pool` to `cylc.scheduler`
for easier access to suite variables.
