+++ doc/siterc.tex
@@ -859,6 +859,8 @@ of which can be found under ~\ref{SuiteEventHandling}:
 
 \subparagraph[mail events]{[cylc] \textrightarrow [[events]] \textrightarrow mail events}
 
+\subparagraph[mail footer]{[cylc] \textrightarrow [[events]] \textrightarrow mail footer}
+
 \subparagraph[mail from]{[cylc] \textrightarrow [[events]] \textrightarrow mail from}
 
 \subparagraph[mail smtp]{[cylc] \textrightarrow [[events]] \textrightarrow mail smtp}
