+++ tests/events/09-task-event-mail.t
@@ -20,11 +20,14 @@
 if ! mail -V 2>'/dev/null'; then
     skip_all '"mail" command not available'
 fi
-set_test_number 5
+set_test_number 3
 mock_smtpd_init
 OPT_SET=
 if [[ "${TEST_NAME_BASE}" == *-globalcfg ]]; then
     create_test_globalrc "" "
+[cylc]
+    [[events]]
+        mail footer = see: http://localhost/stuff/%(owner)s/%(suite)s/
 [task events]
     mail events = failed, retry, succeeded
     mail smtp = ${TEST_SMTPD_HOST}"
@@ -39,9 +42,13 @@ run_ok "${TEST_NAME_BASE}-validate" \
 suite_run_ok "${TEST_NAME_BASE}-run" \
     cylc run --reference-test --debug ${OPT_SET} "${SUITE_NAME}"
 
-grep_ok 'Subject: \[1 task(s) retry\]' "${TEST_SMTPD_LOG}"
-grep_ok 'Subject: \[1 task(s) succeeded\]' "${TEST_SMTPD_LOG}"
-grep_ok '^1/t1/\(01: retry\|02: succeeded\)' "${TEST_SMTPD_LOG}"
+contains_ok "${TEST_SMTPD_LOG}" <<__LOG__
+task 1: 1/t1/01
+task event: retry
+task 1: 1/t1/02
+task event: succeeded
+see: http://localhost/stuff/${USER}/${SUITE_NAME}/
+__LOG__
 
 purge_suite "${SUITE_NAME}"
 mock_smtpd_kill
